Pulse_predictor-main/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
jinja2
pandas
numpy
scikit-learn
joblib
aiofiles
bcrypt==4.0.1
pyxlsb
openpyxl
